render pages - events, thanks, and create #Relates #60 #68
