+++ public/css/style.css
@@ -1,0 +1,6 @@
+body {
+  background-color: #DBB5D5;
+}
+.logo-image {
+  width: 100px;
+}
