+++ src/app.js
@@ -5,6 +5,27 @@ const controllers = require('./controllers/index');
 const helpers = require('./views/helpers/index');
 const app = express();
 
+const mongoose = require('mongoose');
+const mongoDB = 'mongodb://localhost/something-good';
+const db = mongoose.connection;
+
+mongoose.connect(mongoDB);
+
+db.on('error', console.error.bind(console, 'Problem with mongoDB connection'));
+
+const Schema = mongoose.Schema;
+const eventSchema = new Schema({
+  title: String,
+  description: String,
+  location: String,
+  date: Date,
+  image: String,
+  participants: Number
+});
+var eventModel = mongoose.model('eventModel', eventSchema);
+
+module.exports = eventModel;
+
 app.set('views', path.join(__dirname, 'views'));
 app.set('view engine', 'hbs');
 app.engine('hbs',
@@ -16,17 +37,8 @@ exphbs({
   helpers: helpers
 })
 );
-// app.get('/', (req, res) => {
-//   res.render('home', home);
-// });
-//
-// // app.get('/', (req, res) => {
-// //   res.send('Something Good is happening here');
-// // });
-//
-// app.get('/events', (req, res) => {
-//   res.render('events', events);
-// });
+
+app.use(express.static(path.join(__dirname, '..', 'public')));
 
 app.set('port', process.env.PORT || 5000);
