+++ src/controllers/create.js
@@ -1,0 +1,5 @@
+exports.get = (req, res) => {
+  const { create } = req.params;
+  console.log('params for create', req.params);
+  return create.js;
+};
