+++ src/controllers/events.js
@@ -1,5 +1,5 @@
 const events = require('./../../database/db_connection.js');
 
 exports.get = (req, res) => {
-  res.render('events', events);
+  res.render('events', { events });
 };
