+++ src/controllers/home.js
@@ -1,5 +1,5 @@
 exports.get = (req, res) => {
   const { home } = req.params;
-  console.log(req.params);
+  console.log('params for home', req.params);
   return home.js;
 };
