+++ src/controllers/index.js
@@ -2,7 +2,7 @@ const express = require('express');
 const router = express.Router();
 
 const home = require('./');
-// const events = require('./events');
+const events = require('./events');
 // const event = require('./event');
 // const create = require('./create');
 // const thanks = require('./thanks');
@@ -12,12 +12,22 @@ router.get('/', (req, res) => {
   res.render('home', home);
 });
 
+router.get('/events', (req, res) => {
+  res.render('events', events);
+});
+
+router.get('/create', (req, res) => {
+  res.render('create', events);
+});
+
+router.get('/thanks', (req, res) => {
+  res.render('thanks', events);
+});
 // router.get('/', home);
 // router.get('/events', events);
 // router.get('/event', event);
-// router.post('/create', create);
+// router.get('/create', create);
 // router.get('/thanks', thanks);
-
 // router.use(error.client);
 // router.use(error.server);
