+++ src/controllers/thanks.js
@@ -1,0 +1,5 @@
+exports.get = (req, res) => {
+  const { thanks } = req.params;
+  console.log('params for thanks', req.params);
+  return thanks.js;
+};
